// ui/types/css.ts
export enum CSS {
  opaque = 'opaque',
  faded = 'faded',
  hidden = 'hidden',
  hiddenConditionally = 'hidden-conditionally',
  noHide = 'no-hide',

  plistRow = 'plist-row',
  plistRowHighlight = 'plist-row-highlight',

  keyContainer = 'key-container',
  keyContainerLeft = 'key-container-left',
  keyContainerCenter = 'key-container-center',
  keyContainerRight = 'key-container-right',
  keyContainerButton = 'key-container-button',

  expandCollapseButton = 'expand-collapse-button',
  inputAsLabel = 'input-as-label',
  focusedInputAsLabel = 'focused-input-as-label',
  selectAsLabel = 'select-as-label',

  errorFlash = 'error-flash',
}